feat: extract @seans-mfe/contracts workspace package (C1)
Creates packages/contracts/ with all shared types: typed errors,
error classifier, CommandResult envelope, and daemon Message types.
Old src/runtime/errors, error-classifier, and src/oclif/envelope
become one-release-cycle re-export shims.  All in-repo imports
updated to @seans-mfe/contracts.  Adds pnpm-workspace.yaml and
TypeScript paths alias so the package resolves without a publish.

Closes #109
