+++ package.json
@@ -1,9 +1,10 @@
 {
   "name": "seans-mfe-tool",
   "version": "1.0.0",
+  "packageManager": "pnpm@10.33.0",
   "main": "bin/run.js",
   "scripts": {
-    "build": "tsc && node scripts/copy-runtime-files.js && node scripts/remove-dist-shims.js && npx oclif manifest && npm run build:schemas",
+    "build": "npm run build --prefix packages/contracts && tsc && node scripts/copy-runtime-files.js && node scripts/remove-dist-shims.js && npx oclif manifest && npm run build:schemas",
     "build:schemas": "ts-node scripts/generate-schemas.ts",
     "build:schemas:check": "ts-node scripts/generate-schemas.ts --check",
     "build:watch": "tsc --watch",
@@ -70,6 +71,7 @@
     }
   },
   "dependencies": {
+    "@seans-mfe/contracts": "file:packages/contracts",
     "@apidevtools/swagger-parser": "^10.1.0",
     "@babel/generator": "^7.22.1",
     "@babel/parser": "^7.23.0",
