+++ src/mcp/server.ts
@@ -17,7 +17,7 @@ import { spawn } from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import { loadToolRegistry, buildArgv, McpToolDefinition } from './tool-registry';
-import type { CommandResult } from '../oclif/envelope';
+import type { CommandResult } from '@seans-mfe/contracts';
 
 export interface McpServerOptions {
   schemasDir:  string;
