+++ src/runtime/errors/index.ts
@@ -1,12 +1,9 @@
-/**
- * Typed Error Classes
- * Following REQ-LIFECYCLE-005: Error Classification
- * Following ADR-065: Error Classification with Hybrid Detection
- */
-
-export { TimeoutError } from './TimeoutError';
-export { NetworkError } from './NetworkError';
-export { ValidationError } from './ValidationError';
-export { SecurityError } from './SecurityError';
-export { BusinessError } from './BusinessError';
-export { SystemError } from './SystemError';
+// Shim — re-exports from @seans-mfe/contracts for one release cycle.
+export {
+  TimeoutError,
+  NetworkError,
+  ValidationError,
+  SecurityError,
+  BusinessError,
+  SystemError,
+} from '@seans-mfe/contracts';
